+++ extension/src/javascripts/all.js
@@ -2,5 +2,5 @@
 //= require prism-json.js
 //= require prism-autolinker.js
 //= require jsl-format.js
-//= require util.js
+//= require content_util.js
 //= require content.js