+++ extension/src/javascripts/event.js
@@ -1,7 +1,10 @@
+//= require util.js
+//= require_self
+
 chrome.runtime.onMessage.addListener(function(request, sender, sendResponse) {
 
-  if (request.action === "GET_THEME") {
-    sendResponse({theme: localStorage["theme"] || "default"});
+  if (request.action === "GET_OPTIONS") {
+    sendResponse(getOptions());
   }
 
 });