+++ Changelog.md
@@ -4,6 +4,7 @@
   * Six more themes and the ability to change between them
   * bugfix: activate only on json/jsonp sources
   * Better json key parser
+  * Max JSON size option (default: 1024 kb)
 
 # 0.1.0